requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.10
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.8.6
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
swagger-ui-py==22.7.13
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
